updated sbt to 1.3.0-RC5
