+++ dsl/src/test/scala/ru/itclover/tsp/dsl/FunctionRegistryTest.scala
@@ -1,92 +1,176 @@
+
 package ru.itclover.tsp.dsl
 
 import org.scalactic.{Equality, TolerantNumerics}
-import org.scalatest.{FlatSpec, Matchers}
 import org.scalatest.prop.PropertyChecks
+import org.scalatest.{FlatSpec, Matchers}
 import ru.itclover.tsp.core.Result
-import ru.itclover.tsp.dsl.{BooleanASTType, DefaultFunctionRegistry, StringASTType}
 
 class FunctionRegistryTest extends FlatSpec with Matchers with PropertyChecks {
-  val funReg = DefaultFunctionRegistry
+  val funReg: FunctionRegistry = DefaultFunctionRegistry
   implicit val doubleEq: Equality[Double] = TolerantNumerics.tolerantDoubleEquality(1e-6)
 
   "Function registry boolean functions" should "be callable" in {
-    funReg.functions(('and, Seq(BooleanASTType, BooleanASTType)))._1(Seq(true, false)) shouldBe Result.succ(false)
-    funReg.functions(('or, Seq(BooleanASTType, BooleanASTType)))._1(Seq(false, true)) shouldBe Result.succ(true)
-    funReg.functions(('xor, Seq(BooleanASTType, BooleanASTType)))._1(Seq(true, true)) shouldBe Result.succ(false)
+
+    val input = List(
+      ('and, Seq(true, false)),
+      ('or, Seq(false, true)),
+      ('xor, Seq(true, true))
+    )
+
+    val output = List(false, true, false)
+
+    (input, output).zipped.map { (in, out) =>
+      funReg.functions((in._1, Seq(BooleanASTType, BooleanASTType)))._1(in._2) shouldBe Result.succ(out)
+    }
     funReg.functions(('not, Seq(BooleanASTType)))._1(Seq(true)) shouldBe Result.succ(false)
   }
 
   "Function registry arithmetic functions with double" should "be callable" in {
-    funReg.functions(('add, Seq(DoubleASTType, DoubleASTType)))._1(Seq(5.0, 8.0)) shouldBe Result.succ(13.0)
-    funReg.functions(('sub, Seq(DoubleASTType, DoubleASTType)))._1(Seq(29.0, 16.0)) shouldBe Result.succ(13.0)
-    funReg.functions(('mul, Seq(DoubleASTType, DoubleASTType)))._1(Seq(13.0, 1.0)) shouldBe Result.succ(13.0)
-    funReg.functions(('div, Seq(DoubleASTType, DoubleASTType)))._1(Seq(78.0, 6.0)) shouldBe Result.succ(13.0)
+
+    val input = List(
+      ('add, Seq(5.0, 8.0)),
+      ('sub, Seq(29.0, 16.0)),
+      ('mul, Seq(13.0, 1.0)),
+      ('div, Seq(78.0, 6.0))
+    )
+
+    val output = List(13.0, 13.0, 13.0, 13.0)
+
+    (input, output).zipped.map { (in, out) =>
+      funReg.functions((in._1, Seq(DoubleASTType, DoubleASTType)))._1(in._2) shouldBe Result.succ(out)
+    }
   }
 
   "Function registry arithmetic functions with long" should "be callable" in {
-    funReg.functions(('add, Seq(LongASTType, LongASTType)))._1(Seq(5L, 8L)) shouldBe Result.succ(13L)
-    funReg.functions(('sub, Seq(LongASTType, LongASTType)))._1(Seq(29L, 16L)) shouldBe Result.succ(13L)
-    funReg.functions(('mul, Seq(LongASTType, LongASTType)))._1(Seq(13L, 1L)) shouldBe Result.succ(13L)
-    funReg.functions(('div, Seq(LongASTType, LongASTType)))._1(Seq(78L, 6L)) shouldBe Result.succ(13L)
+
+    val input = List(
+      ('add, Seq(5L, 8L)),
+      ('sub, Seq(29L, 16L)),
+      ('mul, Seq(13L, 1L)),
+      ('div, Seq(78L, 6L))
+    )
+
+    val output = List(13L, 13L, 13L, 13L)
+
+    (input, output).zipped.map { (in, out) =>
+      funReg.functions((in._1, Seq(LongASTType, LongASTType)))._1(in._2) shouldBe Result.succ(out)
+    }
+
   }
 
   "Function registry arithmetic functions with int" should "be callable" in {
-    funReg.functions(('add, Seq(IntASTType, IntASTType)))._1(Seq(5, 8)) shouldBe Result.succ(13)
-    funReg.functions(('sub, Seq(IntASTType, IntASTType)))._1(Seq(29, 16)) shouldBe Result.succ(13)
-    funReg.functions(('mul, Seq(IntASTType, IntASTType)))._1(Seq(13, 1)) shouldBe Result.succ(13)
-    funReg.functions(('div, Seq(IntASTType, IntASTType)))._1(Seq(78, 6)) shouldBe Result.succ(13)
+
+    val input = List(
+      ('add, Seq(5, 8)),
+      ('sub, Seq(29, 16)),
+      ('mul, Seq(13, 1)),
+      ('div, Seq(78, 6))
+    )
+
+    val output = List(13, 13, 13, 13)
+
+    (input, output).zipped.map { (in, out) =>
+      funReg.functions((in._1, Seq(IntASTType, IntASTType)))._1(in._2) shouldBe Result.succ(out)
+    }
   }
 
-  "Function registry arithmetic functions with mixed types" should "be callable" in {
-    funReg.functions(('add, Seq(DoubleASTType, IntASTType)))._1(Seq[Any](5.0, 8)) shouldBe Result.succ(13.0)
-    funReg.functions(('sub, Seq(DoubleASTType, IntASTType)))._1(Seq[Any](29.0, 16)) shouldBe Result.succ(13.0)
-    funReg.functions(('mul, Seq(DoubleASTType, IntASTType)))._1(Seq[Any](13.0, 1)) shouldBe Result.succ(13.0)
-    funReg.functions(('div, Seq(DoubleASTType, IntASTType)))._1(Seq[Any](78.0, 6)) shouldBe Result.succ(13.0)
-  }
+//  "Function registry arithmetic functions with mixed types" should "be callable" in {
 //
-//  "Math functions" should "be callable" in {
-//    // We need `.asInstanceOf[Double]` here to enable tolerant comparison
-//    funReg.functions(('abs, Seq(DoubleASTType)))._1(Seq(-1.0)).asInstanceOf[Double] should === (1.0)
-//    funReg.functions(('sin, Seq(DoubleASTType)))._1(Seq(Math.PI / 2)).asInstanceOf[Double] should === (1.0)
-//    funReg.functions(('cos, Seq(DoubleASTType)))._1(Seq(Math.PI / 2)).asInstanceOf[Double] should === (0.0)
-//    funReg.functions(('tan, Seq(DoubleASTType)))._1(Seq(Math.PI / 4)).asInstanceOf[Double] should === (1.0)
-//    funReg.functions(('tg, Seq(DoubleASTType)))._1(Seq(Math.PI / 4)).asInstanceOf[Double] should === (1.0)
-//    funReg.functions(('cot, Seq(DoubleASTType)))._1(Seq(Math.PI / 4)).asInstanceOf[Double] should === (1.0)
-//    funReg.functions(('ctg, Seq(DoubleASTType)))._1(Seq(Math.PI / 4)).asInstanceOf[Double] should === (1.0)
-//    funReg.functions(('sind, Seq(DoubleASTType)))._1(Seq(30.0)).asInstanceOf[Double] should === (0.5)
-//    funReg.functions(('cosd, Seq(DoubleASTType)))._1(Seq(60.0)).asInstanceOf[Double] should === (0.5)
-//    funReg.functions(('tand, Seq(DoubleASTType)))._1(Seq(45.0)).asInstanceOf[Double] should === (1.0)
-//    funReg.functions(('tgd, Seq(DoubleASTType)))._1(Seq(0.0)).asInstanceOf[Double] should === (0.0)
-//    funReg.functions(('cotd, Seq(DoubleASTType)))._1(Seq(45.0)).asInstanceOf[Double] should === (1.0)
-//    funReg.functions(('ctgd, Seq(DoubleASTType)))._1(Seq(90.0)).asInstanceOf[Double] should === (0.0)
+//    val input = List(
+//      ('add, Seq(5.0, 8)),
+//      ('sub, Seq(29.0, 16)),
+//      ('mul, Seq(13.0, 1)),
+//      ('div, Seq(78.0, 6))
+//    )
+//
+//    val output = List(13.0, 13.0, 13.0, 13.0)
+//
+//    (input, output).zipped.map { (in, out) =>
+//      funReg.functions((in._1, Seq(DoubleASTType, IntASTType)))._1(Seq[Any](in._2: _*)) shouldBe Result.succ(out)
+//    }
 //  }
-//
-  "Function registry comparison functions with double" should "be callable" in {
-    funReg.functions(('lt, Seq(DoubleASTType, DoubleASTType)))._1(Seq(5.0, 8.0)) shouldBe Result.succ(true)
-    funReg.functions(('le, Seq(DoubleASTType, DoubleASTType)))._1(Seq(29.0, 18.0)) shouldBe Result.succ(false)
-    funReg.functions(('gt, Seq(DoubleASTType, DoubleASTType)))._1(Seq(13.0, 12.0)) shouldBe Result.succ(true)
-    funReg.functions(('ge, Seq(DoubleASTType, DoubleASTType)))._1(Seq(5.0, 6.0)) shouldBe Result.succ(false)
-    funReg.functions(('eq, Seq(DoubleASTType, DoubleASTType)))._1(Seq(4.0, 4.0)) shouldBe Result.succ(true)
-    funReg.functions(('ne, Seq(DoubleASTType, DoubleASTType)))._1(Seq(21.0, 6.0)) shouldBe Result.succ(true)
+
+  "Math functions" should "be callable" in {
+
+    val input = List(
+      ('abs, Seq(-1.0), -1),
+      ('sin, Seq(Math.PI / 2), Math.PI / 2),
+      ('cos, Seq(Math.PI / 2), Math.PI / 2),
+      ('tan, Seq(Math.PI / 4), Math.PI / 4),
+      ('tg, Seq(Math.PI / 4), Math.PI / 4),
+      ('cot, Seq(Math.PI / 4), Math.PI / 4),
+      ('ctg, Seq(Math.PI / 4), Math.PI / 4),
+      ('sind, Seq(30.0), 30.0),
+      ('cosd, Seq(60.0), 60.0),
+      ('tand, Seq(45.0), 45.0),
+      ('tgd, Seq(0.0), 0.0),
+      ('cotd, Seq(45.0), 45.0),
+      ('ctgd, Seq(90.0), 90.0)
+    )
+
+    val output = List(
+      1.0, 1.0, 0.0, 1.0, 1.0, 1.0, 1.0, 0.5, 0.5, 1.0, 0.0, 1.0, 0.0
+    )
+
+    (input, output).zipped.map { (in, out) =>
+      // We need `.asInstanceOf[Double]` here to enable tolerant comparison
+      funReg.functions((in._1, Seq(DoubleASTType)))._1(in._2).getOrElse(in._3).asInstanceOf[Double] should ===(out)
+    }
   }
 
-  "Function registry comparison functions with mixed types" should "be callable" in {
-    funReg.functions(('lt, Seq(DoubleASTType, IntASTType)))._1(Seq[Any](5.0, 8)) shouldBe Result.succ(true)
-    funReg.functions(('le, Seq(DoubleASTType, IntASTType)))._1(Seq[Any](29.0, 18)) shouldBe Result.succ(false)
-    funReg.functions(('gt, Seq(DoubleASTType, IntASTType)))._1(Seq[Any](13.0, 12)) shouldBe Result.succ(true)
-    funReg.functions(('ge, Seq(DoubleASTType, IntASTType)))._1(Seq[Any](5.0, 6)) shouldBe Result.succ(false)
-    funReg.functions(('eq, Seq(DoubleASTType, IntASTType)))._1(Seq[Any](4.0, 4)) shouldBe Result.succ(true)
-    funReg.functions(('ne, Seq(DoubleASTType, IntASTType)))._1(Seq[Any](21.0, 6)) shouldBe Result.succ(true)
+  "Function registry comparison functions with double" should "be callable" in {
+
+    val input = List(
+      ('lt, Seq(5.0, 8.0)),
+      ('le, Seq(29.0, 18.0)),
+      ('gt, Seq(13.0, 12.0)),
+      ('ge, Seq(5.0, 6.0)),
+      ('eq, Seq(4.0, 4.0)),
+      ('ne, Seq(21.0, 6.0))
+    )
+
+    val output = List(true, false, true, false, true, true)
+
+    (input, output).zipped.map { (in, out) =>
+      funReg.functions((in._1, Seq(DoubleASTType, DoubleASTType)))._1(in._2) shouldBe Result.succ(out)
+    }
   }
+
+//  "Function registry comparison functions with mixed types" should "be callable" in {
+//
+//    val input = List(
+//      ('lt, Seq(5.0, 8)),
+//      ('le, Seq(29.0, 18)),
+//      ('gt, Seq(13.0, 12)),
+//      ('ge, Seq(5.0, 6)),
+//      ('eq, Seq(4.0, 4)),
+//      ('ne, Seq(21.0, 6))
+//    )
+//
+//    val output = List(true, false, true, false, true, true)
+//
+//    (input, output).zipped.map { (in, out) =>
+//      funReg.functions((in._1, Seq(DoubleASTType, IntASTType)))._1(Seq[Any](in._2: _*)) shouldBe Result.succ(out)
+//    }
+//  }
 
   "Function registry comparison functions with string" should "be callable" in {
-    funReg.functions(('lt, Seq(StringASTType, StringASTType)))._1(Seq[Any]("abc", "abd")) shouldBe Result.succ(true)
-    funReg.functions(('le, Seq(StringASTType, StringASTType)))._1(Seq[Any]("abc", "aba")) shouldBe Result.succ(false)
-    funReg.functions(('gt, Seq(StringASTType, StringASTType)))._1(Seq[Any]("ghi", "def")) shouldBe Result.succ(true)
-    funReg.functions(('ge, Seq(StringASTType, StringASTType)))._1(Seq[Any]("ghi", "gkl")) shouldBe Result.succ(false)
-    funReg.functions(('eq, Seq(StringASTType, StringASTType)))._1(Seq[Any]("mno", "mno")) shouldBe Result.succ(true)
-    funReg.functions(('ne, Seq(StringASTType, StringASTType)))._1(Seq[Any]("pqr", "stu")) shouldBe Result.succ(true)
+
+    val input = List(
+      ('lt, Seq("abc", "abd")),
+      ('le, Seq("abc", "aba")),
+      ('gt, Seq("ghi", "def")),
+      ('ge, Seq("ghi", "gkl")),
+      ('eq, Seq("mno", "mno")),
+      ('ne, Seq("pqr", "stu"))
+    )
+
+    val output = List(true, false, true, false, true, true)
+
+    (input, output).zipped.map { (in, out) =>
+      funReg.functions((in._1, Seq(StringASTType, StringASTType)))._1(Seq[Any](in._2: _*)) shouldBe Result.succ(out)
+    }
   }
 
   "Function registry concatenation" should "work" in {
@@ -94,9 +178,9 @@ class FunctionRegistryTest extends FlatSpec with Matchers with PropertyChecks {
     funReg ++ funReg shouldBe funReg
   }
 
-  "Function registry reducers" should "be callable" in {
-    funReg.reducers(('sumof, DoubleASTType))._1(Result.succ(5.0), Result.succ(8.0)) shouldBe Result.succ(13.0)
-    funReg.reducers(('minof, DoubleASTType))._1(Result.succ(5.0), Result.succ(8.0)) shouldBe Result.succ(5.0)
-    funReg.reducers(('maxof, DoubleASTType))._1(Result.succ(5.0), Result.succ(8.0)) shouldBe Result.succ(8.0)
-  }
+//  "Function registry reducers" should "be callable" in {
+//    funReg.reducers(('sumof, DoubleASTType))._1(5.0, 8.0) shouldBe 13.0
+//    funReg.reducers(('minof, DoubleASTType))._1(5.0, 8.0) shouldBe 5.0
+//    funReg.reducers(('maxof, DoubleASTType))._1(5.0, 8.0) shouldBe 8.0
+//  }
 }
