+++ project/build.properties
@@ -1,1 +1,1 @@
-sbt.version=1.3.0-bin-20190814T192004
+sbt.version=1.3.0-RC5
